Fixing up cmake stuff on my end

+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include <execinfo.h>
-#include <signal.h>
+#include <csignal>
 #include <unistd.h>
 
 #include <iostream>
@@ -12,7 +12,7 @@ using namespace std;
 //Shamelessly copied from StackOverflow
 void handler(int sig) {
     void *array[10];
-    size_t size;
+    int size;
 
     // get void*'s for all entries on the stack
     size = backtrace(array, 10);
